Setup build process
The build procedure is as simple as running the .sh file each project folder will contain, first/first.sh in this case. The first line might have to be updated with the path of your devkitARM.
For convenience, you can drop GBA.sublime-build in /Users/setht/Library/Application\ Support/Sublime\ Text\ 3/Packages/User/GBA.sublime-build to let you run the .sh file with ⌘+B. The script also launches an emulator, so with one keystroke your game will build and be running in front of you in just a second.